Abort pgbench if script end is reached with an opened pipeline
When a pipeline is opened with \startpipeline and not closed, pgbench
will either error on the next transaction with a "already in pipeline
mode" error or successfully ends if this was the last transaction
despite not sending anything that was piped in the pipeline.

This change adds an error if the end of script is reached while there's
an opened pipeline.