Add ALL_CANDIDATES explain option
When ALL_CANDIDATES is provided, the query planner is forced to keep all
paths instead of discarding the more expensive option. From those paths,
a list of PlannedStmt is generated which is then outputed by explain.

Due to how planning mutate the PlannerGlobal, we can't directly iterate
the path list generated by the subquery_planner and create planned
statement for them. Instead, we need to plan from scratch for every path
in the pathlist to generate the list of PlannedStmt.