Set query_id for queries contained in utility statement
Some utility statements like Explain, CreateTableAs, DeclareCursor and
CreateMaterializedView contain a query which will be planned and
executed. During post parse, only the top utility statement is jumbled,
leaving the contained query without a set query_id. ExplainQuery does
jumble the other three do not.

This led to extensions relying on query_id like pg_stat_statements to
not be able to track those nested queries as the query_id was 0.

This patch fixes this by jumbling the nested query of CreateTableAs,
and DeclareCursor before it is executed.
Additionally, explain's nested query can itself be a CreateTableAs,
DeclareCursor which also needs to be jumbled. This is now done in
ExplainOneUtility.