Consider pipeline implicit transaction as a transaction block
When using pipeline with implicit transaction, a transaction will
start from the first command and be committed with the Sync message.
Functions like IsInTransactionBlock and PreventInTransactionBlock
already assimilate this implicit transaction as a transaction block.

However, this is not the case in CheckTransactionBlock. This function is
called for things like warning when set local is used outside of a
transaction block.

This patch changes the transaction state to an implicit transaction
block when a new command is processed while pipelining was detected.
This remove the need to check for XACT_FLAGS_PIPELINING in
IsInTransactionBlock and PreventInTransactionBlock since the transaction
state now reflects the ongoing transaction block.

This allows to avoid warning when `set local` is called within a
pipeline, and makes the detection of transaction block coherent with
what's done in IsInTransactionBlock and PreventInTransactionBlock.

A known limitation of this patch: the first command won't be detected as
part of a transaction block. The implicit block only starts at the
second command so the `set local` warnings will still happen if called
first.