Add PgProto test module to send message on a raw socket
Add a new PostgreSQL::Test::PgProto class that provides helper functions
to send PostgreSQL protocol message to a socket. This initial version
provides functions to send a startup packet, send a simple query, read
the content of a response and consume responsed until the socket is
closed.

This is useful to test edge cases like sessions stuck in a specific
socket state.