Accept recovery conflict interrupt on blocked writing
Previously, all interrupts except process dying were ignored while a
process was blocked writing to a socket. If the connection to the client
was broken (no clean FIN nor RST), a process sending results to the
client could be stuck for 924s until the TCP retransmission timeout is
reached. During this time, it was possible for the process to conflict
with recovery: For example, the process returning results can have a
conflicting buffer pin.

To avoid blocking recovery for an extended period of time, this patch
changes client write interrupts by handling recovery conflict interrupts
instead of ignoring them. Since the interrupt happens while we're likely
to have partially written results on the socket, there's no easy way to
keep protocol sync so the session needs to be terminated.

In addition, a blocked write could happen while interrupts are blocked.
For example, we're sending an error message to the client and eventually
saturate the socket buffer. In this case, we can't process the interrupt
in ProcessClientWriteInterrupt. Instead, we stop retrying writes and
trigger a write error, allowing to close the connection and termiate the
session.