Add additional memory context checks
Currently, a deleted aset MemoryContext in the freelist can't be
distinguished from a valid MemoryContext, and will be seen as valid by
checks like MemoryContextIsValid. To fix that, we set the node's type to
T_Invalid before putting it in the freelist, marking the MemoryContext
as deleted and not usable unless it goes through
AllocSetContextCreateInternal first. The T_Invalid node type will allow
to trigger the MemoryContextIsValid and AllocSetIsValid checks.

We also add additional MemoryContextIsValid checks:
- Before restoring the context stored by a
  transactionState->priorContext
- In MemoryContextCreate to make sure the provided parent is valid