Avoid using deleted context with replication command
When creating a new replication slot with snapshot export and using
replication command, a new transaction is started by
SnapBuildExportSnapshot. This transaction will save the memory context
at the start of the transaction, which would be the replication command
context.

Currently, this memory context is deleted at the end of
exec_replication_command. Thus during the next replication command call,
SnapBuildClearExportedSnapshot will abort the previous transaction and
restore the deleted memory context as CurrentMemoryContext.

To fix this issue, the replication command MemoryContext is kept alive
if there's an ongoing snapshot export. This context will be restored
during the next replication command when the snapshot transaction is
aborted.