Fix location and length tracking of select statement
Previously, we tracked the start of a select statement using SELECT's
position. However, this doesn't work for statements like '(SELECT 1)
limit 1;' as the 'limit 1' needs to be part of the statement and thus,
the parenthesis needs to be included.

This patch instead tracks the start of the select_stmt to the outermost
'(' if present. The statement's length is only filled where we can
reliably get the information:
- COPY: We use the position of the '()' surrounding the PreparableStmt
- CTAS: Use the location of the opt_with_data to update the length.

Outside of those cases, the statement length is unset by default and it
will fallback to the RawStmt's length to compute the remaining length.