Rename \close to \close_prepared
\close is a new psql meta-command introduced in d55322b allowing to
close a prepared statement using extended protocol in psql. However, the
name close is ambiguous. At the SQL level, CLOSE is used to close an
existing cursor. At the protocol level, the close message can be used to
close either a statement or a portal.

This patch renames \close to \close_prepared to avoid this ambiguity and
make it clearer that this is used to close a prepared statement.