Doc: improve documentation about stream abort
Protocol v4 introduces parallel streaming. With this change, Stream
Abort messages will have additional abort info when parallel streaming
is enabled. However, the doc only mentions "This field is available
since protocol version 4", which can be confused as they will be
present as long as it's using protocol version 4.

This patch adds precision that abort LSN and TS are only present if
parallel stream is used.